node: add aws.conf to nodes as well
apiserver demands it on non-primary masters
